Move MollieGateway into the new Builtin namespace
