--- src/Gateways/MollieGateway.php
@@ -1,264 +1,0 @@
-<?php
-
-namespace DoubleThreeDigital\SimpleCommerce\Gateways;
-
-use DoubleThreeDigital\SimpleCommerce\Contracts\Gateway;
-use DoubleThreeDigital\SimpleCommerce\Data\Gateways\BaseGateway;
-use DoubleThreeDigital\SimpleCommerce\Data\Gateways\GatewayPrep;
-use DoubleThreeDigital\SimpleCommerce\Data\Gateways\GatewayPurchase;
-use DoubleThreeDigital\SimpleCommerce\Data\Gateways\GatewayResponse;
-use DoubleThreeDigital\SimpleCommerce\Events\PostCheckout;
-use DoubleThreeDigital\SimpleCommerce\Facades\Cart;
-use DoubleThreeDigital\SimpleCommerce\Facades\Currency;
-use DoubleThreeDigital\SimpleCommerce\Facades\Order;
-use DoubleThreeDigital\SimpleCommerce\Facades\Product;
-use DoubleThreeDigital\SimpleCommerce\Facades\Shipping;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Config;
-use Mollie\Api\MollieApiClient;
-use Mollie\Api\Types\PaymentStatus;
-use Statamic\Entries\Entry;
-use Statamic\Facades\Site;
-use Statamic\Facades\Entry as EntryFacade;
-
-class MollieGateway extends BaseGateway implements Gateway
-{
-    protected $mollie;
-
-    public function setup(): void
-    {
-        $this->mollie = new MollieApiClient();
-        $this->mollie->setApiKey($this->config()['key']);
-    }
-
-    public function name(): string
-    {
-        return 'Mollie';
-    }
-
-    public function prepare(GatewayPrep $data): GatewayResponse
-    {
-        $this->setup();
-        $cart = $data->cart();
-
-        if ($this->isUsingPaymentsApi()) {
-            $payment = $this->mollie->payments->create([
-                'amount' => [
-                    'currency' => Currency::get(Site::current())['code'],
-                    'value' => (string) substr_replace($cart->data['grand_total'], '.', -2, 0),
-                ],
-                'description' => "Order {$cart->title}",
-                'redirectUrl' => $this->callbackUrl(),
-                'webhookUrl'  => $this->webhookUrl(),
-                'metadata' => [
-                    'order_id' => $cart->id,
-                ],
-            ]);
-
-            return new GatewayResponse(true, [
-                'id' => $payment->id,
-                'type' => 'payments',
-            ], $payment->getCheckoutUrl());
-        }
-
-        if ($this->isUsingOrdersApi()) {
-            $cart = Order::find($cart->id());
-            $billingAddress = $cart->billingAddress();
-
-            if (! $billingAddress) {
-                throw new \Exception("Please add a billing address to the order before attempting to create a Mollie order.");
-            }
-
-            $order = $this->mollie->orders->create([
-                'amount' => [
-                    'currency' => Currency::get(Site::current())['code'],
-                    'value' => (string) substr_replace($cart->get('grand_total'), '.', -2, 0),
-                ],
-                'orderNumber' => $cart->title,
-                'lines' => $cart->orderItems()
-                    ->map(function ($item) use ($cart) {
-                        $product = Product::find($item['product']);
-                        $taxAmount = $cart->get('tax_total') / $cart->orderItems()->count();
-
-                        if ($product->purchasableType() === 'variants') {
-                            if (is_array($item['variant'])) {
-                                $variant = $product->variantOption($item['variant']['variant']);
-                            } else {
-                                $variant = $product->variantOption($item['variant']);
-                            }
-                        } else {
-                            $variant = null;
-                        }
-
-                        return [
-                            'type' => $product->get('is_digital_product', false) === true
-                                ? 'digital'
-                                : 'physical',
-                            'name' => is_null($variant)
-                                ? $product->title
-                                : "$product->title - {$variant['variant']}",
-                            'quantity' => $item['quantity'],
-                            'unitPrice' => [
-                                'currency' => Currency::get(Site::current())['code'],
-                                'value' => is_null($variant)
-                                    ? (string) substr_replace($product->get('price'), '.', -2, 0)
-                                    : (string) substr_replace($variant['price'], '.', -2, 0)
-                            ],
-                            'totalAmount' => [
-                                'currency' => Currency::get(Site::current())['code'],
-                                'value' => (string) substr_replace($item['total'], '.', -2, 0),
-                            ],
-                            'vatRate' => '0',
-                            'vatAmount' => [
-                                'currency' => Currency::get(Site::current())['code'],
-                                'value' => '0.00'
-                            ],
-                        ];
-                    })
-                    ->merge([
-                        [
-                            'type' => 'shipping_fee',
-                            'name' => 'Shipping',
-                            'quantity' => 1,
-                            'unitPrice' => [
-                                'currency' => Currency::get(Site::current())['code'],
-                                'value' =>  (string) substr_replace($cart->get('shipping_total'), '.', -2, 0),
-                            ],
-                            'totalAmount' => [
-                                'currency' => Currency::get(Site::current())['code'],
-                                'value' =>  (string) substr_replace($cart->get('shipping_total'), '.', -2, 0),
-                            ],
-                            'vatRate' => '0',
-                            'vatAmount' => [
-                                'currency' => Currency::get(Site::current())['code'],
-                                'value' => '0.00',
-                            ],
-                        ],
-                        [
-                            'name' => 'Tax',
-                            'quantity' => 1,
-                            'unitPrice' => [
-                                'currency' => Currency::get(Site::current())['code'],
-                                'value' =>  (string) substr_replace($cart->get('tax_total'), '.', -2, 0),
-                            ],
-                            'totalAmount' => [
-                                'currency' => Currency::get(Site::current())['code'],
-                                'value' =>  (string) substr_replace($cart->get('tax_total'), '.', -2, 0),
-                            ],
-                            'vatRate' => '0',
-                            'vatAmount' => [
-                                'currency' => Currency::get(Site::current())['code'],
-                                'value' => '0.00',
-                            ],
-                        ],
-                    ])
-                    ->toArray(),
-                'billingAddress' => [
-                    'givenName' => $billingAddress->name(), // TODO: this should be first name
-                    'familyName' => $billingAddress->name(), // TODO: this should be last name
-                    'email' => is_null($cart->customer())
-                        ? 'no-email@example.com'
-                        : $cart->customer()->email(),
-                    'streetAndNumber' => $billingAddress->address(),
-                    'postalCode' => $billingAddress->zipCode(),
-                    'city' => $billingAddress->city(),
-                    'country' => $billingAddress->country(),
-                ],
-                'redirectUrl' => $this->callbackUrl(),
-                // 'webhookUrl'  => $this->webhookUrl(),
-                'locale' => 'en_US', // TODO: allow this to be configurable
-            ]);
-
-            return new GatewayResponse(true, [
-                'id' => $order->id,
-                'type' => 'orders',
-            ], $order->getCheckoutUrl());
-        }
-
-        throw new \Exception("Somehow you've got here... You've specified an invalid API type for the Mollie Gateway.");
-    }
-
-    public function purchase(GatewayPurchase $data): GatewayResponse
-    {
-        // We don't actually do anything here as Mollie is an
-        // off-site gateway, so it has it's own checkout page.
-
-        // TODO: maybe throw an exception, in the case a developer gets here?
-
-        return new GatewayResponse(false, []);
-    }
-
-    public function purchaseRules(): array
-    {
-        // Mollie is off-site, therefore doesn't use the traditional
-        // checkout process provided by Simple Commerce. Hence why no rules
-        // are defined here.
-
-        return [];
-    }
-
-    public function getCharge(Entry $order): GatewayResponse
-    {
-        $this->setup();
-        $order = Order::find($order->id());
-
-        $molliePayment = $this->mollie->payments->get(
-            $order->get('gateway_data')['id']
-        );
-
-        return new GatewayResponse(
-            true,
-            (array) $payment
-        );
-    }
-
-    public function refundCharge(Entry $order): GatewayResponse
-    {
-        $this->setup();
-        $order = Order::find($order->id());
-
-        $payment = $this->mollie->payments->get(
-            $cart->get('gateway_data')['id']
-        );
-
-        $refund = $payment->refund([]);
-
-        return new GatewayResponse(
-            true,
-            []
-        );
-    }
-
-    public function webhook(Request $request)
-    {
-        $mollieId = $request->id;
-
-        $payment = $this->mollie->payments->get($mollieId);
-
-        if ($payment->status === PaymentStatus::STATUS_PAID) {
-            $order = Order::query()
-                ->filter(function ($entry) use ($mollieId) {
-                    return isset($entry->data()->get('mollie')['id'])
-                        && $entry->data()->get('mollie')['id']
-                        === $mollieId;
-                })
-                ->map(function ($entry) {
-                    return Cart::find($entry->id());
-                })
-                ->first();
-
-            $order->markAsCompleted();
-            event(new PostCheckout($order->data));
-        }
-    }
-
-    protected function isUsingPaymentsApi()
-    {
-        return $this->configAsCollection()->get('api', 'payments') === 'payments';
-    }
-
-    protected function isUsingOrdersApi()
-    {
-        return $this->configAsCollection()->get('api', 'payments') === 'orders';
-    }
-}
